+++ CHANGELOG.md
@@ -6,6 +6,15 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [v0.2.0] - 2022-06-08
+### Added
+- Warp menu generation
+  - Add runnable to the controller which observes keys in the etcd specified in a configmap `k8s-ces-warp-config`
+  and creates warp menu entries in `k8s-ces-menu-json` for the nginx-ingress dogu [#3].
+
+### Changed
+- Update makefiles to version 6.0.1 [#3]
+
 ## [v0.1.0] - 2022-04-20
 ### Added
 - Automatically creates the ingress class `k8s-ecosystem-ces-service` in the current namespace.